# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TraderMemos API
  version: 1.0.0
  description: |
    Self-hosted trading journal API.

    Authenticate with either:
    - A short-lived session JWT from `POST /api/v1/auth/login` (web app)
    - A personal access token (`tm_pat_…`) from Settings → API or `POST /api/v1/access-tokens`

    Send the secret as `Authorization: Bearer <token>`.
servers:
  - url: /
    description: This API host
...
paths:
...
  /api/v1/trades:
    get:
      tags: [Trades]
      summary: List trades
      operationId: listTrades
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - name: ignored_missing_prices
          in: query
          description: Comma-separated symbol:date pairs explicitly confirmed for previous-close carry-forward
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: target_currency
          in: query
          description: Explicit display currency; returns normalized trades envelope. Native detail and storage remain unchanged.
          schema:
            type: string
      responses:
        "200":
          description: Trades
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      type: object
                  - type: object
                    description: currency, target_currency, fx_policy latest, fx_rates and trades. Row P&L, fees, prices and initial_risk use output currency; source_pnl_currency retains provenance.
                    additionalProperties: true
  /api/v1/trades/{id}:
    get:
      tags: [Trades]
      summary: Get trade
      operationId: getTrade
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Trades]
      summary: Update trade journal fields
      operationId: patchTrade
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated
    delete:
      tags: [Trades]
      summary: Delete trade
      operationId: deleteTrade
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session JWT access token **or** personal access token (`tm_pat_…`).
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    AccountIdQuery:
      name: account_id
      in: query
      description: >-
        Account scope. Accepts a single account id, a comma-separated list, or
        the parameter repeated (portfolio mode). All included accounts must share
        one base currency; mixed-currency selections return 400
        `mixed_currencies`, including when omitted. Omit for all live accounts.
      schema:
        type: string
    FromQuery:
      name: from
      in: query
      description: RFC3339 start time
      schema:
        type: string
        format: date-time
    ToQuery:
      name: to
      in: query
      description: RFC3339 end time
      schema:
        type: string
        format: date-time
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
security:
  - bearerAuth: []